requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Frozen-Flask==0.18
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
pyaml==21.10.1
Pygments==2.15.1
pymdown-extensions==10.0.1
pypandoc==1.11
python-slugify==8.0.1
PyYAML==6.0
requests==2.29.0
text-unidecode==1.3
urllib3==1.26.15
Werkzeug==2.3.3
